+++ courses/tasks.py
@@ -42,9 +42,10 @@ def update_courses_meta_data(*args, **kwargs):
         ),
         "start": course.start and course.start.isoformat(),
         "end": course.end and course.end.isoformat(),
-        "enrollment_start": course.enrollment_start and course.enrollment_start.isoformat(),
+        "enrollment_start": course.enrollment_start
+        and course.enrollment_start.isoformat(),
         "enrollment_end": course.enrollment_end and course.enrollment_end.isoformat(),
-        "languages": ["fr"],
+        "languages": [course.language or "fr"],
     }
     json_data = json.dumps(data)